// src/generated/client/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  INVENTORY_MANAGER: 'INVENTORY_MANAGER',
  EMPLOYEE: 'EMPLOYEE'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const ProductStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  DISCONTINUED: 'DISCONTINUED'
} as const

export type ProductStatus = (typeof ProductStatus)[keyof typeof ProductStatus]


export const MovementType = {
  INCOMING: 'INCOMING',
  OUTGOING: 'OUTGOING',
  ADJUSTMENT: 'ADJUSTMENT'
} as const

export type MovementType = (typeof MovementType)[keyof typeof MovementType]


export const NotificationType = {
  LOW_STOCK: 'LOW_STOCK',
  MOVEMENT: 'MOVEMENT',
  SYSTEM: 'SYSTEM'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]